requirements.txt
llama-cpp-python
distro
termcolor
pyperclip
inquirer
